cli/migrate-to-groups: provide an option to reset an existing group schema
Also do not alter the schema if the group schema has been
detected in the db.

+++ tshistory/schema.py
@@ -13,9 +13,9 @@ class tsschema(object):
     def __init__(self, ns='tsh'):
         self.namespace = ns
 
-    def create(self, engine):
+    def create(self, engine, reset=False):
         self._create_series(engine, self.namespace)
-        self._create_group(engine)
+        self._create_group(engine, reset=reset)
 
     def _create_series(self, engine, namespace):
         with engine.begin() as cn:
@@ -29,8 +29,22 @@ class tsschema(object):
             # creation
             cn.execute(sqlfile(SERIES, ns=namespace))
 
-    def _create_group(self, engine):
+    def _create_group(self, engine, reset=False):
         # cleanup
+        exists = engine.execute(
+            'select 1 from information_schema.schemata where schema_name = %(name)s',
+            name=f'{self.namespace}.group'
+        ).scalar()
+        if exists and not reset:
+            print(f'The "{self.namespace}.group" namespace already exists.')
+            return
+
+        if reset:
+            # undo the contents of group.sql
+            with engine.begin() as cn:
+                cn.execute(f'drop table if exists "{self.namespace}".group_registry')
+                cn.execute(f'drop table if exists "{self.namespace}".groupmap')
+
         with engine.begin() as cn:
             cn.execute(f'drop schema if exists "{self.namespace}.group" cascade')
